requirements.txt
flask==3.0.2
flask-sqlalchemy==3.1.1
flask-login==0.6.3
werkzeug==3.0.1
numpy==1.26.4
scikit-learn==1.4.1
opencv-python==4.9.0.80
Pillow==10.2.0
pandas==2.1.4
xlsxwriter==3.1.9